requirements.txt
fastmcp
pydantic
python-dotenv
requests
uvicorn
